improve calculation accuracy and fix truncation error
The initial change improves the calculation accuracy by allowing the use of float-point arguments instead of integers to avoid truncation errors. A follow-up change fixes the truncation error that occurs when converting float to integer, by introducing rounding instead of truncation.

Test: Add (4.9, 4.9) should return 10.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>

fix truncation error

To convert float to integer the truncation is performed, but the
rounding is expected.

Test: Add (4.9, 4.9) should return 10.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>